backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
PyPDF2
python-dotenv
requests
websockets
openpyxl
python-docx
Pillow
aiofiles
